+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Use `duckdb.sql` and ask only for WHERE clause
 - Update restapi annotations
 - Use `Settings` in restapi/cli core functions
+- Restapi/Cli: Use pydantic models for request parameters
 
 ## 0.98.0 - 2024-12-09